Move AIO into Meta component
